q3

+++ src/se254/a4/q3/A.java
@@ -3,5 +3,5 @@ package se254.a4.q3;
 public class A {
 	public int x;	
 	public double z;
-	public int y;
+	public double y; // test was int now double
 }
